Awjf waldplan (#2099)
* [awjf_waldplan] Sicherung Datenmigraton

* [awjf_waldplan] Korrektur Klassenname

* [awjf_waldplan] new sql-file

* [awjf_waldplan] new sql-files

* [awjf_waldplan] update Datenmigration

* [awjf_waldplan] Update Datenmigration

* [awjf_waldplan] new sql-files

* [awjf_waldplan] update sql

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update Datenmigration

* [awjf_waldplan] update Datenmigration

* [awjf_waldplan] update Datenmigration

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update pub-job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] removed bfsnr-options

* [awjf_waldplan] added jenkinsfile

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] added walduebersicht-sql

* [awjf_waldplan] update job

* [awjf_waldplan] update job

* [awjf_waldplan] changed eigentuemerinformation

* [awjf_waldplan] Korrektur forstbetrieb

* [awjf_waldplan] update forstbetrieb processing-table

* [awjf_waldplan] update job

+++ awjf_waldplan_pub/processing_data/awjf_waldplan_grundstueck_processing.sql
@@ -172,7 +172,7 @@ INSERT INTO grundstuecke
 	LEFT JOIN awjf_waldplan_pub_v2.t_ili2db_basket AS basket
 		ON dataset.t_id = basket.dataset
 	WHERE
-		ww.t_datasetname =  ${bfsnr_param}
+		ww.t_datasetname::int4 = ${bfsnr_param}
 ;
 
 CREATE INDEX 
@@ -191,7 +191,7 @@ INSERT INTO grundstuecke_berechnung
 	LEFT JOIN agi_mopublic_pub.mopublic_grundstueck AS mop
 		ON ww.egrid = mop.egrid
 	WHERE
-		ww.t_datasetname =  ${bfsnr_param}
+		ww.t_datasetname::int4 = ${bfsnr_param}
 	GROUP BY
 		ww.egrid,
 		ww.t_datasetname,
@@ -218,7 +218,7 @@ INSERT INTO waldfunktion
 	FROM 
 		awjf_waldplan_pub_v2.waldplan_waldfunktion AS wf
 	WHERE
-		wf.t_datasetname =  ${bfsnr_param}
+		wf.t_datasetname::int4 = ${bfsnr_param}
 ;
 
 CREATE INDEX 
@@ -236,7 +236,7 @@ INSERT INTO waldnutzung
 	FROM 
 		awjf_waldplan_pub_v2.waldplan_waldnutzung AS wnz
 	WHERE
-		wnz.t_datasetname =  ${bfsnr_param}
+		wnz.t_datasetname::int4 = ${bfsnr_param}
 ;
 
 CREATE INDEX
